refactor: color transfer and mnist (#10)
* feature: color transport dashboard visualization

* refactor: added modules necessary for visualization to pixi

* refactor: color transfer tests

* refactor: mnist minor fixes

* refactor: mnist config update

* refactor: added documentation for color transfer

* refactor: added color transfer docs to index.md

* feature: imporoved color transfer - better problem sampling, rng control, enabled measurement function choice and column drop

* fix: color transfer sampling with no permutations

* fix: mnist magic numbers removed, missing dir behavior

* refactor: mnist tests

* Update README.md

* refactor: updated docs

* refactor: final docs tidy-up

* refactor(color-transfer): reorganize module and enhance logging

* refactor(color-transfer): force float64

* feat(color-transfer): implement ColorTransferProblem and ColorTransferExperiment classes; update image processing and metrics measurement

* refactor(color-transfer): streamline image transformation functions

* fix(measurement): update _wait_jax_finish to use jax.tree.map instead of tree_map

* refactor(measurement): update image processing to use JAX instead of NumPy

* fix(measurement): disable JAX usage in distribution metrics and ensure integer type for indices in pixel transformation

* fix(metrics): ensure src_support and tgt_support are converted to numpy arrays in compute_wasserstein_distance

* fix(color-transfer): added use_cost_matrix option in experiment

* fix(bfm): remove rho_mu_push from metrics

* chore: cleanup old configs and junk files

* config: add synthetic dataset yaml configs

* config: add many runners configs

* fix(measure): update to_discrete method to filter weights based on include_zeros flag

* feat(experiment): add learning rate finder and refactor measurement functions

* fix(generator): adjust mean and variance coefficients for Gaussian Mixture Generator

* fix(pdlp): update marginal handling to use discrete measures directly

* Implement AMSGradSolver for entropic optimal transport with multi-marginal support

- Added AMSGradSolver class for 2-marginal dual gradient ascent using AMSGrad optimization.
- Introduced JIT-compiled core loop for efficient computation of transport plans and costs.
- Refactored gradient ascent logic to support multiple marginals (N=1..3) in GradientAscentMultiMarginalSGD.
- Enhanced error handling for cost shape compatibility with marginals.
- Removed deprecated LogDomainGradientAscentSolver and associated functions.
- Updated SAGA solver to ensure compatibility with discrete measures without including zeros.

* fix(back-and-forth: implement Monge map functionality and refactor gradient computation

- Added `monge_map.py` with functions to compute Monge maps from potential functions in ND.
- Introduced `monge_field_from_psi_nd` to compute the Monge field based on a given potential.
- Implemented `interp_vector_field_at_positions_nd` for multilinear interpolation of vector fields.
- Refactored the gradient computation by moving `_central_gradient_nd` to `uot.utils.central_gradient_nd`.
- Updated `pushforward.py` to utilize the new gradient function and adjusted the computation of `s_raw`.
- Modified `solver.py` to compute the Monge map using the new `monge_map_from_psi_nd` function.
- Cleaned up unused code and comments in `solver.py` related to diagnostics and metrics.

* feat(sinkhorn): add sinkhorn_divergence_with_solver function and normalize cost in SinkhornTwoMarginalLogJaxSolver

* feat(metrics): add various utility functions for transport metrics and computations

* chore(playground): added single py files of solver implementation testing

* chore(notebooks): added examples of benchmarking in notebooks

* fix(benchmark): update timestamp format in export filename to include full date

* feat(scripts): add run_color_transfer script for color transfer benchmarking

* feat(color-transfer): implement sinkhorn divergence computation and update measurement metrics

* chore(.gitignore): add data folders and GPU tracker files to ignore list

* feat(back-and-forth): refactor and enhance the back-and-forth solver and related utilities

- Updated import paths for `_central_gradient_nd` and `cic_pushforward_nd` to improve clarity and organization.
- Implemented a new central gradient function in `pushforward.py` to handle boundary conditions more effectively.
- Enhanced `adaptive_pushforward_nd` to support adaptive sub-cell sampling and improved vertex-based pushforward calculations.
- Added utility functions for computing axis spacings and cell volumes in `generate_nd_grid.py`.
- Improved the `BackNForthSqEuclideanSolver` to allow for flexible pushforward function resolution and added extra metrics for diagnostics.
- Updated the convexity metric function to use modern type hints.
- Minor adjustments to ensure consistent handling of grid shapes and types across various functions.

* fix(experiment): update learning rate selection method and refactor pushforward imports

* chore(notebooks): added sinkhorn sparsity tests

* feat(generators): enhance weight preparation and grid generation in various generator classes

* feat(generators): update dataset count and borders for grid generator configurations

* feat(config): update parameter grids for gradient schedules in adam.yaml

* fix(notebooks): updated some back-and-forth plots

---------

+++ uot/data/dataset_loader.py
@@ -28,11 +28,11 @@ def load_matrix_as_color_grid(pixels: ArrayLike, num_channels: int, bins_per_cha
         multipliers = np.array([bins_per_channel ** i for i in reversed(range(num_channels))])
         flat_bins = (bins @ multipliers).astype(np.int32)
 
-        weights_1d = jnp.ones(flat_bins.shape[0], dtype=jnp.float32)
+        weights_1d = jnp.ones(flat_bins.shape[0], dtype=jnp.float64)
         flat_hist = segment_sum(weights_1d, flat_bins, bins_per_channel ** num_channels)
         weights_nd = flat_hist.reshape([bins_per_channel] * num_channels)
     else:
-        weights_nd = np.zeros([bins_per_channel] * num_channels, dtype=np.float32)
+        weights_nd = np.zeros([bins_per_channel] * num_channels, dtype=np.float64)
         for idx in bins:
             weights_nd[tuple(idx)] += 1
 
@@ -59,7 +59,7 @@ def load_image_as_color_grid(
         data = data[:, :, None]
 
     num_channels = data.shape[2]
-    pixels = data.reshape(-1, num_channels).astype(lib.float32) / 255.0
+    pixels = data.reshape(-1, num_channels).astype(lib.float64) / 255.0
 
     return load_matrix_as_color_grid(
         pixels=pixels,
